requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=9.0.0
numpy>=1.24.0
pycocotools>=2.0.6
opencv-python>=4.7.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
tqdm>=4.65.0
matplotlib>=3.7.0
